@@ -1,14 +1,12 @@
 {
   "name": "netbrothers-gmbh/syncacc-bundle",
   "description": "Using NetBrothers Access Control Center in Symfony",
-  "version": "0.0.2",
   "keywords": ["symfony","bundle","permissions","acl"],
   "type": "symfony-bundle",
   "require": {
-    "php": ">=7.4",
+    "php": ">=8.1",
     "symfony/console": "*",
     "symfony/dotenv": "*",
-    "symfony/orm-pack": "*",
     "ext-json": "*"
   },
   "autoload": {
